requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
appdirs==1.4.4
APScheduler==3.9.1
argos-translate-files==1.2.0
argostranslate==1.9.6
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.9.3
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.2
commonmark==0.9.1
cryptography==44.0.0
ctranslate2==4.5.0
Deprecated==1.2.15
discord.py==2.4.0
expiringdict==1.2.2
filelock==3.16.1
Flask==2.2.5
flask-babel==3.1.0
Flask-Limiter==2.6.3
Flask-Session==0.4.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.4.2
langdetect==1.0.9
LexiLang==1.0.1
limits==3.14.1
lit==18.1.8
lxml==5.3.0
MarkupSafe==3.0.2
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.2
pip-autoremove==0.10.0
pip-licenses==5.0.0
pipdeptree==2.24.0
polib==1.1.1
prettytable==3.12.0
prometheus-client==0.15.0
propcache==0.2.1
protobuf==5.29.2
pur==7.3.3
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==4.4.4
regex==2024.11.6
requests==2.31.0
rich==12.6.0
sacremoses==0.0.53
safetensors==0.4.5
sentencepiece==0.2.0
six==1.17.0
soupsieve==2.6
stanza==1.1.1
sympy==1.13.3
tokenizers==0.21.0
tomli==2.2.1
torch==2.0.1
tqdm==4.67.1
transformers==4.47.1
translatehtml==1.5.2
triton==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==2.3.8
wrapt==1.17.0
yarl==1.18.3
